Week_2/docker-requirements.txt
pandas
prefect-gcp[cloud_storage]
protobuf
pyarrow==10.0.1
